Update projects instead of create
There seems to be some sort of issue regarding the cache.  Even if the
cache already contains projects, it doesn't seem to use that existing
table, instead it wants to create a new table.  Even though we are
checking if projects is not an empty return before creating the table.

We need to add debug messages throughout the application to find out
potentially where issues are.